Cleanup python source files
