+++ python/hpsmc/__init__.py
@@ -1,4 +1,4 @@
-from util import config_logging
+from .util import config_logging
 
 # Enable default logging config
 config_logging()
