+++ python/hpsmc/component.py
@@ -2,19 +2,11 @@
 
 import os
 import subprocess
-import sys
-import shutil
-import argparse
-import getpass
-import json
-#import logging
-import time
-import re
+import logging
 
 from util import convert_config_value
 
-# Logging disabled; job should be used instead
-#logger = logging.getLogger("hpsmc.component")
+logger = logging.getLogger("hpsmc.component")
 
 class Component(object):
     """Base class for components in a job."""
@@ -86,7 +78,7 @@ class Component(object):
     def cmd_exists(self):
         """Check if the component's assigned command exists."""
         return subprocess.call("type " + self.command, shell=True,
-            stdout=subprocess.PIPE, stderr=subprocess.PIPE) == 0
+            stdout = subprocess.PIPE, stderr=subprocess.PIPE) == 0
 
     def cmd_args(self):
         """Return the command arguments of this component."""
@@ -115,10 +107,10 @@ class Component(object):
         if parser.has_section(section_name):
             for name, value in parser.items(section_name):
                 setattr(self, name, convert_config_value(value))
-                #logger.info("%s:%s:%s=%s" % (self.name,
-                #                             name,
-                #                             getattr(self, name).__class__.__name__,
-                #                             getattr(self, name)))
+                logger.debug("%s:%s:%s=%s" % (self.name,
+                                             name,
+                                             getattr(self, name).__class__.__name__,
+                                             getattr(self, name)))
 
     def set_parameters(self, params):
         """Set class attributes for the component based on JSON parameters.
@@ -134,20 +126,19 @@ class Component(object):
             else:
                 if p not in self.ignore_job_params:
                     setattr(self, p, params[p])
-                    #logger.info("%s:%s=%s [required]" % (self.name, p, params[p]))
-                #else:
-                #    logger.info("Ignored job param '%s'" % p)
+                    logger.debug("%s:%s=%s [required]" % (self.name, p, params[p]))
+                else:
+                    logger.debug("Ignored job param '%s'" % p)
 
         # Set optional parameters.
         for p in self.optional_parameters():
             if p in params:
                 if p not in self.ignore_job_params:
                     setattr(self, p, params[p])
-                    #logger.info("%s:%s=%s [optional]"
-                    #            % (self.name, p, params[p]))
+                    logger.debug("%s:%s=%s [optional]"
+                                % (self.name, p, params[p]))
                 else:
-                    pass
-                    #logger.info("Ignored job param '%s'" % p)
+                    logger.debug("Ignored job param '%s'" % p)
 
     def required_parameters(self):
         """Return a list of required parameters.
@@ -207,10 +198,10 @@ class Component(object):
         """Configure component from environment variables which are just upper case
         versions of the required config names set in the shell environment."""
         for c in self.required_config():
-            #logger.debug("Setting config '%s' from environ" % c)
+            logger.debug("Setting config '%s' from environ" % c)
             if c.upper() in os.environ:
                 setattr(self, c, os.environ[c.upper()])
-                #logger.debug("Set config '%s=%s' from env var '%s'" % (c, getattr(self, c), c.upper()))
+                logger.debug("Set config '%s=%s' from env var '%s'" % (c, getattr(self, c), c.upper()))
             else:
                 raise Exception("Missing config in environ for '%s'" % c)
